(SIMP-4923) Support rsync server and client
This set of changes allows the ability to have a server support being a
SIMP rsync server and a SIMP rsync client of another server
simultaneously. One instance where this is valuable, is to support a PE
Compile Master <-> Master of Masters host configuration scenario.

These changes combine the rsync::server and rsync::server::global
classes. Global was marked private and is not be used separately in any
other SIMP modules; however, added a significant amount of complexity to
the configuration to support this scenario.

Acceptance tests were added that test being a server and client both
with and without stunnel.

SIMP-4923 #close #comment Support simultaneous server and client